// ticket-app/src/app/types/projectTypes.ts
export type ticketType = {
  origin: string
  origin_name: string
  destination: string
  destination_name: string
  departure_date: string
  departure_time: string
  arrival_date: string
  arrival_time: string
  carrier: string
  stops: number
  price: number
  priceUSD?: number
  priceEUR?: number
}

export type stopsType = {
  checkNum: number
  isChecked: boolean
}